LangGraph/requirements.txt
#langgrapgh
langgrapgh


#langchain
langchain
langchain-google-genai
langchain-community

#additional 
ddgs
bs4
async
await
playwright
yfinance
nest_asyncio
streamlit
langgraph-checkpoint-sqlite
